+++ EconomySimulator2/SupplyMonthly.cs
@@ -9,16 +9,16 @@ namespace EconomySimulator2
 
         public double[] fluct;
 
-        public SupplyMonthly(Good good, double supplyBase, double stockbase, double currentstock, double[] fluctuation)
+        public SupplyMonthly(Good good, int supplyBase, int stockbase, int currentstock, double[] fluctuation)
             : base(good, supplyBase, stockbase, currentstock)
         {
             this.fluct = fluctuation;
         }
 
-        public override double GetBaseSupply(int time)
+        public override int GetBaseSupply(int time)
         {
             int month = (time % fluct.Length);
-            return fluct[month] * supplyBase;
+            return (int)(fluct[month] * supplyBase);
         }
     }
